chore: update environment variables a service locator was added
