+++ .env.template
@@ -1,0 +1,2 @@
+envSupabaseDbUrl = 
+apiKeySupabase =
