+++ lib/config/config.dart
@@ -1,0 +1,8 @@
+export 'package:web_portfolio/config/constants/colors.dart';
+export 'package:web_portfolio/config/constants/enviroment.dart';
+export 'package:web_portfolio/config/constants/nav_items.dart';
+export 'package:web_portfolio/config/constants/size.dart';
+export 'package:web_portfolio/config/constants/skill_items.dart';
+export 'package:web_portfolio/config/theme/theme.dart';
+
+export 'package:web_portfolio/config/data_dev.dart';
