+++ lib/config/data_dev.dart
@@ -8,7 +8,4 @@ class Dev {
   static String urlBlog = '';
 
   static String flutterVersion = '3.22.3';
-  static String envSupabaseDbUrl = 'https://kfazkyriwrlnetfvqkcz.supabase.co';
-  static String apiKeySupabase =
-      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImtmYXpreXJpd3JsbmV0ZnZxa2N6Iiwicm9sZSI6ImFub24iLCJpYXQiOjE3MjEzNDM5ODQsImV4cCI6MjAzNjkxOTk4NH0.uQ4tJ3QCztr6j1V1iwk4kW4cD5bNygkYaIJJAoU3dCk';
 }
