+++ lib/features/home/presentation/widgets/drawer_mobile.dart
@@ -1,0 +1,42 @@
+import 'package:web_portfolio/features/home/presentation/screens.dart';
+
+class DrawerMobile extends StatelessWidget {
+  const DrawerMobile({super.key, required this.onNavItemTap});
+  final Function(int) onNavItemTap;
+
+  @override
+  Widget build(BuildContext context) {
+    return Drawer(
+      backgroundColor: CustomColor.scaffoldBg,
+      child: ListView(
+        children: [
+          Align(
+            alignment: Alignment.centerLeft,
+            child: Padding(
+              padding: const EdgeInsets.only(left: 20, top: 20, bottom: 20),
+              child: IconButton(
+                  onPressed: () {
+                    Navigator.of(context)
+                        .pop(); //TODO: Se debe migrar a GoRouter
+                  },
+                  icon: const Icon(Icons.close)),
+            ),
+          ),
+          for (int i = 0; i < navIcons.length; i++)
+            ListTile(
+              contentPadding: const EdgeInsets.symmetric(horizontal: 30.0),
+              leading: Icon(navIcons[i]),
+              title: Text(navTitles[i]),
+              titleTextStyle: const TextStyle(
+                  color: CustomColor.whitePrimary,
+                  fontWeight: FontWeight.w600,
+                  fontSize: 16),
+              onTap: () {
+                onNavItemTap(i);
+              },
+            )
+        ],
+      ),
+    );
+  }
+}
