+++ lib/features/home/presentation/widgets/footer_section/footer.dart
@@ -1,0 +1,19 @@
+import 'package:web_portfolio/features/home/presentation/screens.dart';
+
+class Footer extends StatelessWidget {
+  const Footer({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.symmetric(vertical: 20),
+      width: double.maxFinite,
+      alignment: Alignment.center,
+      child: Text(
+        'Made by ${Dev.name} with Flutter ${Dev.flutterVersion}',
+        style: const TextStyle(
+            fontWeight: FontWeight.w400, color: CustomColor.whiteSecondary),
+      ),
+    );
+  }
+}
