+++ lib/features/home/presentation/widgets/header_section/header_web_desktop.dart
@@ -1,0 +1,44 @@
+import 'package:web_portfolio/features/home/presentation/screens.dart';
+
+class HeaderDesktop extends StatelessWidget {
+  const HeaderDesktop({super.key, required this.onNavMenuTap});
+  final Function(int) onNavMenuTap;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 60,
+      margin: const EdgeInsets.symmetric(
+        vertical: 10.0,
+        horizontal: 20.0,
+      ),
+      width: double.maxFinite,
+      decoration: BoxDecoration(
+        gradient: const LinearGradient(
+            colors: [Colors.transparent, CustomColor.bgLight1]),
+        borderRadius: BorderRadius.circular(100),
+      ),
+      child: Row(
+        children: [
+          SiteLogo(
+            onTap: () {},
+          ),
+          const Spacer(),
+          for (int i = 0; i < navTitles.length; i++)
+            Padding(
+              padding: const EdgeInsets.only(right: 20),
+              child: TextButton(
+                  onPressed: () {
+                    onNavMenuTap(i);
+                  },
+                  child: Text(
+                    navTitles[i],
+                    style: const TextStyle(
+                        fontSize: 16, color: CustomColor.whitePrimary),
+                  )),
+            ),
+        ],
+      ),
+    );
+  }
+}
