+++ lib/features/home/presentation/widgets/site_logo.dart
@@ -1,0 +1,22 @@
+import 'package:web_portfolio/features/home/presentation/screens.dart';
+
+class SiteLogo extends StatelessWidget {
+  const SiteLogo({super.key, this.onTap});
+
+  final VoidCallback? onTap;
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: () {},
+      child: const Text(
+        'Flutter Dev',
+        style: TextStyle(
+            fontSize: 22,
+            fontWeight: FontWeight.bold,
+            decoration: TextDecoration.underline,
+            color: CustomColor.yellowSecondary),
+      ),
+    );
+  }
+}
