+++ lib/features/home/presentation/widgets/styles/style.dart
@@ -1,0 +1,7 @@
+import 'package:web_portfolio/features/home/presentation/screens.dart';
+
+BoxDecoration kHeaderDecoration = BoxDecoration(
+  gradient:
+      const LinearGradient(colors: [Colors.transparent, CustomColor.bgLight1]),
+  borderRadius: BorderRadius.circular(100),
+);
