+++ lib/features/shared/presentation/widgets/loadings/custom_loading.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class CustomLoading extends StatelessWidget {
+  const CustomLoading({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(child: CircularProgressIndicator());
+  }
+}
