--- lib/presentation/screens.dart
@@ -1,26 +1,0 @@
-// Flutter packages
-export 'package:flutter/material.dart';
-
-// Flutter team packages
-export 'package:web_portfolio/config/data_dev.dart';
-export 'package:web_portfolio/config/theme.dart';
-export 'package:web_portfolio/presentation/pages/home_pages.dart';
-export 'package:web_portfolio/config/constants/colors.dart';
-export 'package:web_portfolio/config/constants/size.dart';
-export 'package:web_portfolio/presentation/widgets/contact_section/contact_section.dart';
-export 'package:web_portfolio/presentation/widgets/drawer_mobile.dart';
-export 'package:web_portfolio/presentation/widgets/footer_section/footer.dart';
-export 'package:web_portfolio/presentation/widgets/header_section/header_mobile.dart';
-export 'package:web_portfolio/presentation/widgets/header_section/header_web_desktop.dart';
-export 'package:web_portfolio/presentation/widgets/main_section/main_desktop.dart';
-export 'package:web_portfolio/presentation/widgets/main_section/main_mobile.dart';
-export 'package:web_portfolio/presentation/widgets/project_skill_section/projects_section.dart';
-export 'package:web_portfolio/presentation/widgets/project_skill_section/skills_desktop.dart';
-export 'package:web_portfolio/presentation/widgets/project_skill_section/skills_mobile.dart';
-export 'package:web_portfolio/presentation/widgets/custom_text_field.dart';
-export 'package:web_portfolio/presentation/widgets/site_logo.dart';
-export 'package:web_portfolio/presentation/widgets/styles/style.dart';
-export 'package:web_portfolio/config/constants/nav_items.dart';
-export 'package:web_portfolio/domain/project_utils.dart';
-export 'package:web_portfolio/presentation/widgets/project_skill_section/project_card.dart';
-export 'package:web_portfolio/config/constants/skill_items.dart';
