--- lib/presentation/widgets/custom_text_field.dart
@@ -1,39 +1,0 @@
-import 'package:web_portfolio/presentation/screens.dart';
-
-class CustomTextField extends StatelessWidget {
-  const CustomTextField(
-      {super.key,
-      this.controller,
-      this.maxLines = 1, //by defualt
-      this.hintText});
-
-  //receive properties to use in different places
-  //TODO: Riverpod to manage this states valors
-  final TextEditingController? controller;
-  final int maxLines;
-  final String? hintText;
-
-  @override
-  Widget build(BuildContext context) {
-    return TextField(
-      controller: controller,
-      maxLines: maxLines,
-      style: const TextStyle(
-        color: CustomColor.scaffoldBg,
-      ),
-      decoration: InputDecoration(
-          contentPadding: const EdgeInsets.all(16),
-          filled: true,
-          fillColor: CustomColor.whiteSecondary,
-          enabledBorder: getInputBorder,
-          border: getInputBorder,
-          hintText: hintText,
-          hintStyle: const TextStyle(color: CustomColor.hintDark)),
-    );
-  }
-
-  OutlineInputBorder get getInputBorder {
-    return OutlineInputBorder(
-        borderRadius: BorderRadius.circular(10), borderSide: BorderSide.none);
-  }
-}
