--- lib/presentation/widgets/footer_section/footer.dart
@@ -1,19 +1,0 @@
-import 'package:web_portfolio/presentation/screens.dart';
-
-class Footer extends StatelessWidget {
-  const Footer({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.symmetric(vertical: 20),
-      width: double.maxFinite,
-      alignment: Alignment.center,
-      child: Text(
-        'Made by ${Dev.name} with Flutter ${Dev.flutterVersion}',
-        style: const TextStyle(
-            fontWeight: FontWeight.w400, color: CustomColor.whiteSecondary),
-      ),
-    );
-  }
-}
