--- lib/presentation/widgets/project_skill_section/skills_desktop.dart
@@ -1,63 +1,0 @@
-import 'package:web_portfolio/presentation/screens.dart';
-
-class SkillsDesktop extends StatelessWidget {
-  const SkillsDesktop({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        //platforms
-        ConstrainedBox(
-          constraints: const BoxConstraints(maxWidth: 450),
-          child: Wrap(spacing: 5.0, runSpacing: 5.0, children: [
-            for (int i = 0; i < platformItems.length; i++)
-              Container(
-                  width: 200,
-                  decoration: BoxDecoration(
-                      color: CustomColor.bgLight2,
-                      borderRadius: BorderRadius.circular(5)),
-                  child: ListTile(
-                    contentPadding: const EdgeInsets.symmetric(
-                        horizontal: 20.0, vertical: 10.0),
-                    leading: Image.asset(
-                      platformItems[i]["img"],
-                      width: 26.0,
-                    ),
-                    title: Text(platformItems[i]["title"]),
-                  )),
-          ]),
-        ),
-
-        const SizedBox(
-          width: 50.0,
-        ),
-
-        //Skills
-
-        Flexible(
-          child: ConstrainedBox(
-            constraints: const BoxConstraints(maxWidth: 500),
-            child: Wrap(
-              spacing: 10.0,
-              runSpacing: 10.0,
-              children: [
-                for (int i = 0; i < skillItems.length; i++)
-                  Chip(
-                      padding: const EdgeInsets.symmetric(
-                        vertical: 12.0,
-                        horizontal: 16.0,
-                      ),
-                      backgroundColor: CustomColor.bgLight2,
-                      label: Text(skillItems[i]["title"]),
-                      avatar: Image.asset(skillItems[i]["img"]))
-              ],
-            ),
-          ),
-        )
-      ],
-    );
-  }
-}
