--- lib/presentation/widgets/styles/style.dart
@@ -1,7 +1,0 @@
-import 'package:web_portfolio/presentation/screens.dart';
-
-BoxDecoration kHeaderDecoration = BoxDecoration(
-  gradient:
-      const LinearGradient(colors: [Colors.transparent, CustomColor.bgLight1]),
-  borderRadius: BorderRadius.circular(100),
-);
